issue-30557 新增/创建/合同产生账单 增加数据来源区分

+++ ehrest/src/main/java/com/everhomes/rest/asset/AssetPaymentBillSourceId.java
@@ -9,19 +9,19 @@ import com.everhomes.util.StringHelper;
  * </ul>
  */
 public enum AssetPaymentBillSourceId {
-    CREATE((byte)1), IMPORT((byte)2);
+    CREATE(1L), IMPORT(2L);
 
-    private Byte code;
+    private Long code;
 
-    private AssetPaymentBillSourceId(Byte code){
+    private AssetPaymentBillSourceId(Long code){
         this.code = code;
     }
 
-    public Byte getCode() {
+    public Long getCode() {
         return code;
     }
 
-    public static AssetPaymentBillSourceId fromCode(Byte code) {
+    public static AssetPaymentBillSourceId fromCode(Long code) {
         for(AssetPaymentBillSourceId v : AssetPaymentBillSourceId.values()) {
             if(v.getCode() == code)
                 return v;
